Update not to use name utils since Python conflicts with that; remove some outdated tests
